requirements.txt
googletrans==4.0.0rc1
Kivy==2.1.0
openai==0.27.2
PyPDF2==3.0.1
PyQt5==5.15.9
python-dotenv==1.0.0
